+++ src/Data/MailCollection.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Xammie\Mailbook;
+namespace Xammie\Mailbook\Data;
 
 use Illuminate\Support\Collection;
-use Xammie\Mailbook\Data\MailableItem;
 
 class MailCollection
 {
